added failing test for async transform
when using async transform, if the input data comes faster
than the async process can deal with it, the stringifier
is used instead. This is because line 259 of index.coffee
says to use the stringifier if still tranforming and because
the async transformer has not called the callback yet it is
still transforming. This results in a bunch of empty rows
followed by the header (if used) and a single line of output.